added python notebooks to tokenize stickers
