--- .wrangler/tmp/bundle-zNybV1/middleware-insertion-facade.js
@@ -1,3 +1,0 @@
-import { __facade_registerInternal__ } from "E:\\DOCUMENTS\\sticker-compose\\sticker-composer\\node_modules\\wrangler\\templates\\middleware\\loader-sw.ts";
-import * as __MIDDLEWARE_0__ from "E:\\DOCUMENTS\\sticker-compose\\sticker-composer\\node_modules\\wrangler\\templates\\middleware\\middleware-miniflare3-json-error.ts";
-__facade_registerInternal__([__MIDDLEWARE_0__.default])
