+++ player_list.json
@@ -1,0 +1,402 @@
+[
+  "abE",
+  "ableJ",
+  "acoR",
+  "AdreN",
+  "adreN",
+  "advent",
+  "Aerial",
+  "aizy",
+  "Aleksib",
+  "ALEX",
+  "aliStair",
+  "allu",
+  "almazer",
+  "AmaNEk",
+  "ANDROID",
+  "ANGE1",
+  "ANNIHILATION",
+  "apEX",
+  "arT",
+  "arya",
+  "Attacker",
+  "aumaN",
+  "autimatic",
+  "Ax1Le",
+  "AZR",
+  "B1ad3",
+  "b1t",
+  "balblna",
+  "Bart4k",
+  "biguzera",
+  "BIT",
+  "blameF",
+  "bLitz",
+  "BnTeT",
+  "bodyy",
+  "boltz",
+  "bondik",
+  "Boombl4",
+  "BOROS",
+  "Brehze",
+  "broky",
+  "Brollan",
+  "BUDA",
+  "buster",
+  "byali",
+  "cadiaN",
+  "cajunb",
+  "Calyx",
+  "captainMo",
+  "CeRq",
+  "chelo",
+  "chopper",
+  "chrisJ",
+  "COLDYY1",
+  "coldzera",
+  "CRUC1AL",
+  "crush",
+  "Cypher",
+  "daps",
+  "dav1d",
+  "DavCost",
+  "DAVEY",
+  "DD",
+  "DeadFox",
+  "degster",
+  "DemQQ",
+  "denis",
+  "dennis",
+  "dephh",
+  "DeStiNy",
+  "device",
+  "DEVIL",
+  "devoduvek",
+  "dexter",
+  "dgt",
+  "DickStacy",
+  "Dima",
+  "dimasick",
+  "disco doplan",
+  "Dosia",
+  "draken",
+  "drop",
+  "dumau",
+  "dupreeh",
+  "Dycha",
+  "Edward",
+  "electronic",
+  "EliGE",
+  "emagine",
+  "erkaSt",
+  "es3tag",
+  "EspiranTo",
+  "Ethan",
+  "Ex6TenZ",
+  "exit",
+  "f0rest",
+  "F1KU",
+  "facecrack",
+  "FalleN",
+  "fame",
+  "FaNg",
+  "Farlig",
+  "FASHR",
+  "faveN",
+  "felps",
+  "fer",
+  "fitch",
+  "FL1T",
+  "FlameZ",
+  "flamie",
+  "floppy",
+  "flusha",
+  "FNS",
+  "fnx",
+  "Forester",
+  "fox",
+  "freakazoid",
+  "Freeman",
+  "friberg",
+  "frozen",
+  "FugLy",
+  "Furlan",
+  "gade",
+  "GeT_RiGhT",
+  "gla1ve",
+  "gob b",
+  "Golden",
+  "Goofy",
+  "Gratisfaction",
+  "Grim",
+  "GruBy",
+  "GuardiaN",
+  "gxx-",
+  "hades",
+  "hallzerk",
+  "hampus",
+  "Happy",
+  "hardzao",
+  "hatz",
+  "Havoc",
+  "hazed",
+  "headtr1ck",
+  "HEN1",
+  "HexT",
+  "Hiko",
+  "History",
+  "Hobbit",
+  "hooch",
+  "HooXi",
+  "HS",
+  "huNter-",
+  "hutji",
+  "Hyper",
+  "iM",
+  "imoRR",
+  "innocent",
+  "INS",
+  "interz",
+  "isak",
+  "ISSAA",
+  "jabbi",
+  "JaCkz",
+  "Jame",
+  "James",
+  "jasonR",
+  "JDC",
+  "jdm64",
+  "Jerry",
+  "jkaem",
+  "jks",
+  "jL",
+  "jmqa",
+  "JOTA",
+  "jR",
+  "JT",
+  "juanflatroo",
+  "JUGi",
+  "junior",
+  "JW",
+  "k0nfig",
+  "k1to",
+  "kabal",
+  "karrigan",
+  "Karsa",
+  "Kaze",
+  "keev",
+  "KEi",
+  "kennyS",
+  "KENSi",
+  "Keoz",
+  "keshandr",
+  "kinqie",
+  "kioShiMa",
+  "Kjaerbye",
+  "kNgV-",
+  "koosta",
+  "Krad",
+  "kRaSnaL",
+  "Krimbo",
+  "KRIMZ",
+  "KrizzeN",
+  "kRYSTAL",
+  "KSCERATO",
+  "Kvik",
+  "Kylar",
+  "Kyojin",
+  "kyxsan",
+  "latto",
+  "lauNX",
+  "LEGIJA",
+  "Lekr0",
+  "LETN1",
+  "Liazz",
+  "LNZ",
+  "LoveYY",
+  "loWel",
+  "Lucaozy",
+  "LUCAS1",
+  "Lucky",
+  "luken",
+  "m0NESY",
+  "maden",
+  "Magisk",
+  "magixx",
+  "Maikelele",
+  "MAJ3R",
+  "malta",
+  "Maniac",
+  "markeloff",
+  "max",
+  "mezii",
+  "MICHU",
+  "mir",
+  "misutaaa",
+  "mixwell",
+  "mou",
+  "MSL",
+  "mynio",
+  "n0rb3r7",
+  "n0thing",
+  "NAF",
+  "nafany",
+  "nawwk",
+  "NBK-",
+  "neaLaN",
+  "NEKiZ",
+  "NEO",
+  "NEOFRAG",
+  "NertZ",
+  "nex",
+  "nexa",
+  "ngiN",
+  "nicoodoz",
+  "Nifty",
+  "NiKo",
+  "niko",
+  "nin9",
+  "nitr0",
+  "Norwi",
+  "npl",
+  "NQZ",
+  "oBo",
+  "olofmeister",
+  "oSee",
+  "oskar",
+  "ottoNd",
+  "pashaBiceps",
+  "Patsi",
+  "paz",
+  "peet",
+  "Perfecto",
+  "Pimp",
+  "pita",
+  "Plopski",
+  "Professor_Chaos",
+  "pronax",
+  "pyth",
+  "qikert",
+  "r3salt",
+  "rain",
+  "rallen",
+  "Ramz1kBO$$",
+  "refrezh",
+  "reltuC",
+  "REZ",
+  "Rickeh",
+  "rigoN",
+  "roeJ",
+  "ropz",
+  "rox",
+  "RpK",
+  "RUBINO",
+  "RUSH",
+  "S0tF1k",
+  "s1mple",
+  "S1ren",
+  "saffee",
+  "SANJI",
+  "ScreaM",
+  "sdy",
+  "seang@res",
+  "seized",
+  "SENER1",
+  "sergej",
+  "sh1ro",
+  "ShahZaM",
+  "shalfey",
+  "Shara",
+  "shox",
+  "shroud",
+  "SicK",
+  "Sico",
+  "sinnopsyy",
+  "siuhy",
+  "SIXER",
+  "sjuush",
+  "sk0R",
+  "Skadoodle",
+  "skullz",
+  "slaxz-",
+  "SmithZz",
+  "smooya",
+  "Snappi",
+  "snatchie",
+  "Snax",
+  "SnypeR",
+  "somebody",
+  "spaze",
+  "speed4k",
+  "Spiidi",
+  "Spinx",
+  "SPUNJ",
+  "Staehr",
+  "stanislaw",
+  "stavn",
+  "steel",
+  "sterling",
+  "Stewie2K",
+  "STYKO",
+  "Summer",
+  "suNny",
+  "SunPayus",
+  "svyat",
+  "syrsoN",
+  "t0rick",
+  "tabseN",
+  "TACO",
+  "tarik",
+  "TaZ",
+  "Techno4K",
+  "tenzki",
+  "TeSeS",
+  "Thomas",
+  "THREAT",
+  "tiziaN",
+  "tonyblack",
+  "torzsi",
+  "TRY",
+  "Tuurtle",
+  "Twist",
+  "Twistzz",
+  "USTILO",
+  "v$m",
+  "v4lde",
+  "vexite",
+  "vice",
+  "VINI",
+  "volt",
+  "w0nderful",
+  "waterfaLLZ",
+  "wayLander",
+  "WOOD7",
+  "WorldEdit",
+  "Woro2k",
+  "woxic",
+  "xand",
+  "XANTARES",
+  "xccurate",
+  "xertioN",
+  "xfl0ud",
+  "Xizt",
+  "xms",
+  "xsepower",
+  "xseveN",
+  "Xyp9x",
+  "yam",
+  "yay",
+  "YEKINDAR",
+  "yel",
+  "yuurih",
+  "zehN",
+  "Zero",
+  "Zeus",
+  "zevy",
+  "zhokiNg",
+  "znajder",
+  "zorte",
+  "Zyphon",
+  "ZywOo",
+  "alexibob"
+]
